26.07.10 실습1-1. 가격 계산 시 double로 자동 업캐스팅이 되기 때문에 명시적으로 다운캐스팅을 통해 소수점 자리 제거

+++ src/main/java/com/example/demo/PaymentPrice.java
@@ -13,7 +13,7 @@ public class PaymentPrice extends OriginalPrice {
 
     @Override
     public void display() {
-        System.out.printf("상품(%s)의 가격은 %s 입니다.%n", super.name, super.price * TAX_RATE + this.deliveryPrice);
+        System.out.printf("상품(%s)의 가격은 %s 입니다.%n", super.name, (int) (super.price * TAX_RATE + this.deliveryPrice));
     }
